milestone_3/requirement.txt
pandas==2.3.3        
numpy                
seaborn             
matplotlib          
plotly               
prophet==1.2.1       
scikit-learn==1.7.2  
python-dotenv        
nltk                
wordcloud            
newsapi-python==0.2.7  
transformers==4.57.1   
google-generativeai==0.8.5 
requests 